chore: Refatorado a forma de resend das mensagens

+++ src/main/java/com/hav/imobiliaria/controller/dto/usuario/UsuarioListaSelectResponseDTO.java
@@ -1,7 +1,7 @@
 package com.hav.imobiliaria.controller.dto.usuario;
 
 public record UsuarioListaSelectResponseDTO(
-         Integer id,
-         String nome)
+        Long id,
+        String nome)
 {
 }
